Test unit for LCIndexReader
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@144 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Frameworks/LuceneKit/Headers/Index/LCDocumentWriter.h
@@ -16,7 +16,7 @@
 }
 
 - (id) initWithTerm: (LCTerm *) t
-       position: (int) position
+       position: (long) position
        offset: (LCTermVectorOffsetInfo *) offset;
 - (LCTerm *) term;
 - (long) freq;
@@ -64,7 +64,7 @@
 - (void) invertDocument: (LCDocument *) doc;
 - (void) addField: (NSString *) field
              text: (NSString *) text
-         position: (int) position
+         position: (long) position
 	 offset: (LCTermVectorOffsetInfo *) offset;
 - (NSArray *) sortPostingTable;
 #if 0
